Display Account Details
Unable to test it but I'm sure it works

+++ lib/core/account/widgets/card_info.dart
@@ -13,7 +13,7 @@ class AccountCardInfo extends StatefulWidget {
   final String lastName;
   final String cardType;
 
-  final int currAmount;
+  final double currAmount;
 
   AccountCardInfo(
       {required this.accountType,
